use conditional exports
