+++ packages/data-uri-to-buffer/package.json
@@ -4,6 +4,10 @@
   "description": "Create an ArrayBuffer instance from a Data URI string",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
+  "exports": {
+    "node": "./dist/node.js",
+    "default": "./dist/index.js"
+  },
   "files": [
     "dist"
   ],
