+++ packages/data-uri-to-buffer/test/data-uri-to-buffer.test.ts
@@ -1,19 +1,11 @@
 import assert from 'assert';
-import { dataUriToBuffer } from '../src';
+import { dataUriToBuffer as baseline } from '../src/index';
+import { dataUriToBuffer as node } from '../src/index';
 
-describe('data-uri-to-buffer', function () {
-	const useNativeDecodingInTest = () => {
-		delete (globalThis as Record<string, unknown>).__useCustomDecodeInTests;
-	};
+describe('node', () => doTest(node));
+describe('baseline', () => doTest(baseline));
 
-	beforeEach(() => {
-		(globalThis as Record<string, unknown>).__useCustomDecodeInTests = true;
-	});
-
-	afterEach(() => {
-		useNativeDecodingInTest();
-	});
-
+function doTest(dataUriToBuffer: typeof baseline) {
 	it('should decode bare-bones Data URIs', function () {
 		const uri = 'data:,Hello%2C%20World!';
 
@@ -25,16 +17,6 @@ describe('data-uri-to-buffer', function () {
 	});
 
 	it('should decode bare-bones "base64" Data URIs', function () {
-		const uri = 'data:text/plain;base64,SGVsbG8sIFdvcmxkIQ%3D%3D';
-
-		const parsed = dataUriToBuffer(uri);
-		assert.equal('text/plain', parsed.type);
-		assert.equal('Hello, World!', Buffer.from(parsed.buffer).toString());
-	});
-
-	it('should decode bare-bones "base64" Data URIs using native path', function () {
-		useNativeDecodingInTest();
-
 		const uri = 'data:text/plain;base64,SGVsbG8sIFdvcmxkIQ%3D%3D';
 
 		const parsed = dataUriToBuffer(uri);
@@ -209,4 +191,4 @@ describe('data-uri-to-buffer', function () {
 		assert.equal('UTF-8', parsed.charset);
 		assert.equal('abc', Buffer.from(parsed.buffer).toString());
 	});
-});
+}
